New folder/requirements.txt
discord-py
groq
python-dotenv